Convert DndResources to TypeScript
